# Conflicts:
#	bundles/AdminBundle/Resources/public/js/pimcore/object/tags/manyToManyObjectRelation.js

+++ models/DataObject/Fieldcollection/Dao.php
@@ -80,23 +80,34 @@ class Dao extends Model\Dao\AbstractDao
                         if (!$fd instanceof CustomResourcePersistingInterface) {
                             Tool::triggerMissingInterfaceDeprecation(get_class($fd), 'load', CustomResourcePersistingInterface::class);
                         }
-                        // datafield has it's own loader
-                        $value = $fd->load(
-                            $collection,
-                            [
-                                'context' => [
-                                    'object' => $object,
-                                    'containerType' => 'fieldcollection',
-                                    'containerKey' => $type,
-                                    'fieldname' => $this->model->getFieldname(),
-                                    'index' => $result['index']
-                            ]]
-                        );
-                        if ($value === 0 || !empty($value)) {
-                            $collection->setValue($key, $value);
 
-                            if ($collection instanceof DataObject\DirtyIndicatorInterface) {
-                                $collection->markFieldDirty($key, false);
+                        $doLoad = true;
+                        if ($fd instanceof DataObject\ClassDefinition\Data\Relations\AbstractRelations) {
+                            if (!DataObject\Concrete::isLazyLoadingDisabled() && $fd->getLazyLoading()) {
+                                $doLoad = false;
+                            }
+                        }
+
+                        if ($doLoad) {
+                            // datafield has it's own loader
+                            $value = $fd->load(
+                                $collection,
+                                [
+                                    'context' => [
+                                        'object' => $object,
+                                        'containerType' => 'fieldcollection',
+                                        'containerKey' => $type,
+                                        'fieldname' => $this->model->getFieldname(),
+                                        'index' => $result['index']
+                                    ]]
+                            );
+
+                            if ($value === 0 || !empty($value)) {
+                                $collection->setValue($key, $value);
+
+                                if ($collection instanceof DataObject\DirtyIndicatorInterface) {
+                                    $collection->markFieldDirty($key, false);
+                                }
                             }
                         }
                     }
@@ -212,8 +223,8 @@ class Dao extends Model\Dao\AbstractDao
             }
         }
 
-        if (!$this->model->isFieldDirty('_self')) {
-            return false;
+        if (!$this->model->isFieldDirty('_self') && !DataObject\AbstractObject::isDirtyDetectionDisabled()) {
+            return [];
         }
         $whereLocalizedFields = "(ownertype = 'localizedfield' AND "
             . $this->db->quoteInto('ownername LIKE ?', '/fieldcollection~'
@@ -221,11 +232,11 @@ class Dao extends Model\Dao\AbstractDao
             . ' AND ' . $this->db->quoteInto('src_id = ?', $object->getId()). ')';
 
         if ($saveMode) {
-            if (!$this->model->hasDirtyFields() && $hasLocalizedFields) {
+            if (!DataObject\AbstractObject::isDirtyDetectionDisabled() && !$this->model->hasDirtyFields() && $hasLocalizedFields) {
                 // always empty localized fields
                 $this->db->deleteWhere('object_relations_' . $object->getClassId(), $whereLocalizedFields);
 
-                return false;
+                return ['saveLocalizedRelations' => true];
             }
         }
 
@@ -236,6 +247,6 @@ class Dao extends Model\Dao\AbstractDao
         // empty relation table
         $this->db->deleteWhere('object_relations_' . $object->getClassId(), $where);
 
-        return true;
+        return ['saveFieldcollectionRelations' => true, 'saveLocalizedRelations' => true];
     }
 }
